+++ src/components/history/NoDataFound.tsx
@@ -1,0 +1,7 @@
+import { Heading } from '../common';
+
+const NoDataFound = () => {
+  return <Heading level={6}>🗑️ 아직 아무것도 없네요!</Heading>;
+};
+
+export default NoDataFound;
